rename buildtool filenames.
