+++ src/build-tool/common.ts
